add plain script to test noriTokenController burn
